+++ fuser-app/compile.sh
@@ -1,0 +1,2 @@
+rm -rf fs/App.js 
+tsc --module amd --moduleResolution node --outFile js/App.js src/App.ts